+++ Twilio/Rest/Api/V2010/Account/Sip/CredentialList/CredentialContext.php
@@ -10,6 +10,7 @@
 namespace Twilio\Rest\Api\V2010\Account\Sip\CredentialList;
 
 use Twilio\InstanceContext;
+use Twilio\Options;
 use Twilio\Values;
 use Twilio\Version;
 
@@ -62,14 +63,14 @@ class CredentialContext extends InstanceContext {
     /**
      * Update the CredentialInstance
      * 
-     * @param string $username The username
-     * @param string $password The password
+     * @param array|Options $options Optional Arguments
      * @return CredentialInstance Updated CredentialInstance
      */
-    public function update($username, $password) {
+    public function update($options = array()) {
+        $options = new Values($options);
+        
         $data = Values::of(array(
-            'Username' => $username,
-            'Password' => $password,
+            'Password' => $options['password'],
         ));
         
         $payload = $this->version->update(